separated styles for lyrics
