+++ src/screens/LyricsToScreen.tsx
@@ -1,48 +1,45 @@
 /* eslint-disable prefer-const */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 // Worked on by: Ashley Bickham and Siri Avula
-
 import React, { useState, useEffect } from "react";
-import {
-  StyleSheet,
-  Text,
-  TouchableHighlight,
-  View,
-  PixelRatio,
-  Dimensions,
-  ScrollView,
-} from "react-native";
+import { Text, TouchableHighlight, View, ScrollView } from "react-native";
 import UserReaderWriter from "../services/UserReaderWriter";
 import LyricsService from "../services/LyricsService";
 import TranslationService from "../services/TranslationService";
 import WordModal from "../components/WordModal";
+import lyricsStyles from "../styles/LyricsStyles";
 
-const fontScale = PixelRatio.getFontScale();
-const getFontSize = (size) => size / fontScale;
-
-// saves word that was pressed's data, needs to be loaded here so data doesn't get rewritten while data gets updated from API calls
-
-const { height: SCREEN_HEIGHT } = Dimensions.get("window");
-
-// page contents, runs through on each render of page
 export default function LyricsToScreen({ currentTrack }) {
-  //grabbing Song information from song playback
   const playlistItem = currentTrack;
-  // songLang determined by examining song lyrics with detect later
-  //pref language determined by detecting language from user's prefLanguage data
-  /*locally held prefLang, assigned value from userPrefLang value
-  Is a separate value because the formatting of the language is a a two-character code 
-  vs plain english (database pref language format).
-  Some function renders are reliant when this value changes
-  and we do not want those functions to run with the impropely formatted database userPrefLang value*/
   const [prefLang, setPrefLang] = useState("");
-
   const [lyrics, setLyrics] = useState("");
   const [translation, setTranslation] = useState("");
   const [fromLang, setFromLang] = useState("");
-
   const [openModal, setOpenModal] = React.useState(false);
   const [clickedWord, setClickedWord] = React.useState("");
+
+  async function getUserPrefLang() {
+    await UserReaderWriter.getPreferredLanguage().then((DBPrefLang) => {
+      if (DBPrefLang.toLowerCase() == "english") {
+        setPrefLang("en");
+      } else if (
+        DBPrefLang.toLowerCase() == "deutsch" ||
+        DBPrefLang.toLowerCase() == "german"
+      ) {
+        setPrefLang("de");
+      } else if (
+        DBPrefLang.toLowerCase() == "espanol" ||
+        DBPrefLang.toLowerCase() == "spanish"
+      ) {
+        setPrefLang("es");
+      } else if (
+        DBPrefLang.toLowerCase() == "francias" ||
+        DBPrefLang.toLowerCase() == "french"
+      ) {
+        setPrefLang("fr");
+      }
+    });
+  }
 
   async function getLyrics() {
     await LyricsService.getLyrics(playlistItem).then((lyricsResponse) => {
@@ -66,13 +63,10 @@ export default function LyricsToScreen({ currentTrack }) {
   }, []);
 
   return (
-    <View style={styles.container}>
+    <View style={lyricsStyles.container}>
       <ScrollView>
         {lyrics.split("\n").map((line, lineIdx) => (
-          <View
-            key={lineIdx}
-            style={{ marginBottom: 10, flexDirection: "row", flexWrap: "wrap" }}
-          >
+          <View key={lineIdx} style={lyricsStyles.lineFormat}>
             {line.split(" ").map((word, wordIdx) => (
               <TouchableHighlight
                 key={wordIdx}
@@ -80,10 +74,10 @@ export default function LyricsToScreen({ currentTrack }) {
                   setClickedWord(word);
                   setOpenModal(true);
                 }}
-                style={{ marginRight: 4, marginBottom: 4 }}
+                style={lyricsStyles.highlight}
                 underlayColor="#5bc8a6"
               >
-                <Text style={styles.lyricsText}>{word}</Text>
+                <Text style={lyricsStyles.lyricsText}>{word}</Text>
               </TouchableHighlight>
             ))}
           </View>
@@ -91,10 +85,7 @@ export default function LyricsToScreen({ currentTrack }) {
         <Text>TRANSLATION</Text>
 
         {translation.split("\n").map((line, lineIdx) => (
-          <View
-            key={lineIdx}
-            style={{ marginBottom: 10, flexDirection: "row", flexWrap: "wrap" }}
-          >
+          <View key={lineIdx} style={lyricsStyles.lineFormat}>
             {line.split(" ").map((word, wordIdx) => (
               <TouchableHighlight
                 key={wordIdx}
@@ -102,10 +93,10 @@ export default function LyricsToScreen({ currentTrack }) {
                   setClickedWord(word);
                   setOpenModal(true);
                 }}
-                style={{ marginRight: 4, marginBottom: 4 }}
+                style={lyricsStyles.highlight}
                 underlayColor="#5bc8a6"
               >
-                <Text style={styles.lyricsText}>{word}</Text>
+                <Text style={lyricsStyles.lyricsText}>{word}</Text>
               </TouchableHighlight>
             ))}
           </View>
@@ -113,7 +104,7 @@ export default function LyricsToScreen({ currentTrack }) {
 
         <Text>{"\n\n"}</Text>
       </ScrollView>
-      <View style={styles.saveWordModal}>
+      <View style={lyricsStyles.saveWordModal}>
         {openModal && (
           <WordModal
             openModal={openModal}
@@ -126,114 +117,4 @@ export default function LyricsToScreen({ currentTrack }) {
       </View>
     </View>
   );
-
-  async function getUserPrefLang() {
-    await UserReaderWriter.getPreferredLanguage().then((DBPrefLang) => {
-      if (DBPrefLang.toLowerCase() == "english") {
-        setPrefLang("en");
-      } else if (
-        DBPrefLang.toLowerCase() == "deutsch" ||
-        DBPrefLang.toLowerCase() == "german"
-      ) {
-        setPrefLang("de");
-      } else if (
-        DBPrefLang.toLowerCase() == "espanol" ||
-        DBPrefLang.toLowerCase() == "spanish"
-      ) {
-        setPrefLang("es");
-      } else if (
-        DBPrefLang.toLowerCase() == "francias" ||
-        DBPrefLang.toLowerCase() == "french"
-      ) {
-        setPrefLang("fr");
-      }
-      console.log("CODE IS: " + prefLang);
-    });
-  }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    alignItems: "center",
-    backgroundColor: "#303248",
-    paddingTop: 20,
-    height: SCREEN_HEIGHT - 50,
-  },
-  UIInfo: {
-    fontSize: getFontSize(25),
-    color: "#edc526",
-    textAlign: "center",
-  },
-  space: {
-    height: 40,
-    backgroundColor: "white",
-  },
-  noteText: {
-    fontSize: getFontSize(15),
-    color: "#ff4a2a",
-    textAlign: "left",
-    padding: 15,
-  },
-  lyricBlock: {
-    padding: 32,
-    paddingTop: 0,
-    lineHeight: 45,
-    flexWrap: "wrap",
-    flexDirection: "row",
-    color: "white",
-  },
-  lyricsText: {
-    fontSize: getFontSize(25),
-    color: "white",
-  },
-  noFlex: {
-    flexWrap: "nowrap",
-  },
-  buttons: {
-    alignContent: "flex-end",
-    color: "white",
-    fontWeight: "bold",
-    fontSize: 20,
-  },
-  modalbg: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "rgba(0,0,0,0.5)",
-  },
-  modalforefront: {
-    backgroundColor: "#303248",
-    padding: 15,
-    borderRadius: 15,
-    borderWidth: 10,
-    borderColor: "#5bc8a6",
-  },
-  modalText: {
-    color: "white",
-    fontSize: getFontSize(25),
-  },
-  closeModal: {},
-  saveWordModal: {
-    zIndex: -10,
-  },
-  modalButtons: {
-    alignContent: "flex-end",
-    color: "white",
-    fontWeight: "bold",
-    fontSize: 20,
-    marginHorizontal: 5,
-  },
-  placeholderStyle: {
-    fontSize: getFontSize(20),
-    color: "white",
-  },
-  selectedTextStyle: {
-    fontSize: getFontSize(20),
-    color: "white",
-  },
-  iconStyle: {
-    width: 20,
-    height: 20,
-    color: "white",
-  },
-});
